WotCon: extract pure helpers from WotConNodeManager
Move the JSON-LD parser, the JSON->BuiltInType mapper, the mock value generator, and the WotAsset model out of the partial NodeManager into dedicated files so the manager focuses on address-space wiring and OPC UA service handlers.

- ThingDescriptionParser.cs: ThingDescriptionInfo + ThingPropertyInfo + static Parse(json, logger) + GetBuiltInTypeFromJson(jsonType). PropertyInfo renamed to ThingPropertyInfo to avoid System.Reflection.PropertyInfo shadowing.

- WotMockValueGenerator.cs: static Generate(dataTypeId) for the Double/Int32/Boolean/String defaults used when materializing simulated property values.

- WotAsset.cs: per-asset runtime model (Name, AssetId, ThingDescription, FileNodeId, FileMethodMap, FileBuffers, FileLock, NextFileHandle, LastFinalizedPayload).

- WotConNodeManager.cs: redirect CreateAssetNode call sites to the extracted helpers, drop the three private classes, drop the now-unused System.Text.Json import. No behavioral change.

Verified: dotnet build clean, all 11 WotCon tests pass, image redeployed to test-e2e.
